reorganize file structure of application to allow view of github pages. file structure is alot simpler this way
