+++ deploy.sh
@@ -4,10 +4,10 @@
 set -e
 
 # build
-npm run browser-build
+npm run dist
 
 # navigate into the build output directory
-cd browser-build
+cd dist
 
 # if you are deploying to a custom domain
 # echo 'www.example.com' > CNAME
