+++ index.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+    <style>
+        html {
+            color: white;
+            background-color: #282c34;
+        }
+        a {
+            color: white;
+        }
+    </style>
+</head>
+
+<body>
+    <p>
+        This is the testing browser for notepad.
+    </p>
+    <p>Click the links below</p>
+    <a href="notepad.html">Notepad</a>
+    <a href="preferences.html">Preferences</a>
+</body>
+
+</html>
