+++ notepad.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="./main.jsx"></script>
+    <script type="module" src="./src/notepad.jsx"></script>
   </body>
 </html>
