+++ src-tauri/tauri.conf.json
@@ -4,7 +4,7 @@
     "version": "0.2.0"
   },
   "build": {
-    "distDir": "../build",
+    "distDir": "../dist",
     "devPath": "http://localhost:5173",
     "beforeDevCommand": "npm run dev",
     "beforeBuildCommand": "npm run build"
