+++ src/notepad.jsx
@@ -1,12 +1,11 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './App'
+import Notepad from './pages/Notepad'
 
 const container = document.getElementById('root')
 const root = createRoot(container)
 
 root.render(
   <React.StrictMode>
-    <App />
+    <Notepad />
   </React.StrictMode>)
