+++ src/pages/Notepad/Notepad.css
@@ -2,6 +2,14 @@ html {
   background-color: black;
   /* overflow: hidden; */
 }
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
 .paper {
   padding: 0;
   display: block;
