+++ src/preferences.jsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './App'
+import Preferences from './pages/Preferences'
 
 const container = document.getElementById('root')
 const root = createRoot(container)
 
 root.render(
   <React.StrictMode>
-    <App />
+    <Preferences />
   </React.StrictMode>)
