--- src/preferences/App.css
@@ -1,12 +1,0 @@
-#preferences {
-    padding: 10px;
-    height: auto;
-    display: block;
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    background-color: #282c34;
-    color:white;
-}
